+++ docs/changes.md
@@ -5,9 +5,13 @@ title: Model Change Log
 
 ## Version 3.2 {#v3-2}
 
+### v3.2.2 {#v3-2-2}
+
+* Add `getMetaData()` and `setMetaData()` to `Orchestra\Model\Traits\Metable`.
+
 ### v3.2.1 {#v3-2-1}
 
-* Remove `Trait` prefix and deprecate old traits.
+* Remove `Trait` suffix and deprecate old traits.
 * Add `Orchestra\Model\Traits\Searchable`.
 
 ### v3.2.0 {#v3-2-0}